backend-razorpay-6-7-24/requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
zope.interface==6.3
